major credential re-design
git-svn-id: https://helianto.svn.sourceforge.net/svnroot/helianto/trunk@208 d46e4f78-7810-0410-b419-9f95c2a9a517

+++ helianto/core/src/main/java/org/helianto/core/service/CoreMgr.java
@@ -87,7 +87,7 @@ public interface CoreMgr extends GenericService {
      * Return the last available number to a number type
      * and assign a new one with increment 1.
      */
-    public int currentNumber(Entity entity, String type);
+    public long currentNumber(Entity entity, String type);
     
     /**
      * Send a <code>Credential</code> registration using
